estructured the packages, add factory class

+++ src/slogo/model/backEndInternal/CommandExecutor.java
@@ -14,9 +14,9 @@ public class CommandExecutor implements Executor {
      *
      */
     public Object executeCommand(Command currentCommand) throws ExecutionException {
-        //System.out.println("So far good"+currentCommand.execute().toString());
+       //System.out.println("So far good"+currentCommand.execute().toString());
 
-       return currentCommand.execute();
+       return (Double) currentCommand.execute();
     }
